Finish content for carne asada file and add link
Finished adding image, description, ingredients, and steps
content for carne asada tacos recipe. Furthermore, a link the the
webpage was also added to index.html in order to make sure the site
is cohesive and accessible. Creating the content of the page
further reinforced learning about images, links, headings, and lists.