Finish last recipe file and change file name
The lasagna.html file was renamed to pot-pie.html due to a change in
recipe. Then, all of the content for this file was then completed
in order to complete the criteria for the project. Please make another
commit sometime that credits the original articles at the bottom of
the page before uploading.