Delete lasagna file and add acknowledgments
Old recipe file (lasagna.html) was deleted since this file is not
relevant as pot-pie.html was used as a substitute recipe. Moreover,
acknowledgments to sources used for the content of each webpage
were created at the bottom of each page. This is to ensure no plagiarism
takes place if submitted. As of now, this is the last revision needed
for the assignment "Projects:Recipes" in the foundations course of the
Odin Project.