Fill in information for README.md file
Filled in information about project in the README.md file.
Discussed basic discription of project, skills utilized, and
thoughts/conclusion on this part of the project. Completing this file
was imperative for providing context for the project as well as for
completing all criteria specified in the Odin Project lesson
description.